client_template/requirements.txt
aiohttp
datapizza-ai
python-dotenv